Need Help with the Blur Effect
I used pictures as background of section tag before, now, I created a div class and contained the two pictures( org and blur one).
But now the text missed in the home page, I tried to reposition it or adjust the index but doesn't work, and it is also have an influence on the background of "practice" section.  Weird.